+++ src/styles/blocks/slider.scss
@@ -1,0 +1,117 @@
+.about {
+  overflow: hidden;
+
+  &__slider {
+    position: relative;
+    height: max-content;
+
+    margin-top: 20px;
+
+    @include onTablet {
+      grid-column: 1 / 4;
+      margin-top: 0;
+      width: 100%;
+
+      top: 50%;
+      transform: translateY(-50%);
+    }
+
+    @include onDesktop {
+      grid-column: 3 / 8;
+    }
+
+    @include onLarge {
+      grid-column: 3 / 11;
+    }
+  }
+
+  &__slider-nav-buttons {
+    display: flex;
+    justify-content: space-between;
+
+    width: 210px;
+
+    @include onDesktop {
+      margin-top: 12px;
+    }
+  }
+
+  &__slider-nav-buttons-button {
+    display: none;
+
+    font-weight: 500;
+    color: inherit;
+
+    background: none;
+    outline: none;
+    border: none;
+
+    transition: all $animationDuration;
+    cursor: pointer;
+
+    &:hover {
+      color: $contrastColor;
+    }
+
+    @include onDesktop {
+      display: block;
+    }
+  }
+
+  &__slider-nav-buttons-next {
+    color: $white;
+  }
+
+  &__slider-fraction {
+    display: none;
+
+    color: $contrastColor;
+
+    @include onDesktop {
+      display: flex;
+
+      position: absolute;
+      top: -5px;
+      left: -40px;
+    }
+  }
+}
+
+.swiper {
+  &-pagination {
+    transform: translateY(30px);
+
+    @include onDesktop {
+      display: none;
+    }
+  }
+
+  &-pagination-bullet {
+    width: 4px;
+    height: 4px;
+
+    background-color: $white;
+  }
+
+  &-scrollbar {
+    display: none;
+
+    bottom: 5px !important;
+    left: 0 !important;
+
+    width: 210px !important;
+    height: 1px !important;
+
+    background-color: $ytButtonColor;
+    transform: translateY(25px);
+
+    @include onDesktop {
+      display: block;
+    }
+  }
+
+  &-scrollbar-drag {
+    height: 1px;
+    background-color: $contrastColor;
+  }
+}
