+++ src/styles/main.scss
@@ -7,8 +7,11 @@
 @import "./blocks/page.scss";
 
 @import "./blocks/button.scss";
+@import "./blocks/video.scss";
+@import "./blocks/slider.scss";
 
 @import "./blocks/testimonials.scss";
+@import "./blocks/about.scss";
 @import "./blocks/tech.scss";
 @import "./blocks/whyKatLoco.scss";
 @import "./blocks/contact.scss";
