+++ src/styles/utils/_variables.scss
@@ -1,14 +1,20 @@
-$contrastColor: #05c2df;
+$black: #000;
 $backgroundBlack: #05040b;
 $darkGray: #2f2f2f;
-$textColor: #929292;
+
 $white: #fff;
-$gradient: linear-gradient(100.83deg, #191536 0.67%, #000 100%);
-$animationDuration: 0.3s;
+$textColor: #929292;
+$contrastColor: #05c2df;
+$brightRed: #f00;
+$red: #860404;
+
 $buttonActive: #191536;
 $buttonActiveText: #2f2f2f;
-$black: #000;
 $buttonBorderColor: rgba(255, 255, 255, 0.14);
-$buttonPressedBackgroundColor: #07798b;
 $buttonMessageColor: #191536;
-$red: #860404;
+$buttonPressedBackgroundColor: #07798b;
+
+$ytButtonColor: #212121;
+
+$gradient: linear-gradient(100.83deg, #191536 0.67%, #000 100%);
+$animationDuration: 0.3s;
